cli/tests/requirements.txt
aiohttp==3.7.4.post0
black==20.8b1
coverage==5.5
flake8==3.8.4
flake8-quotes==3.2.0
foxglove-web==0.0.18;python_version>="3.8"
isort==5.7.0
mypy==0.812
pycodestyle==2.6.0
pyflakes==2.2.0
pytest==6.2.2
pytest-cov==2.11.1
pytest-mock==3.5.1
pytest-sugar==0.9.4
twine==3.3.0
